split two function from utils to make StudpidEmoji small
