+++ src/utils/clear.js
@@ -1,0 +1,38 @@
+export function unifiedToNative(unified) {
+  var unicodes = unified.split('-'),
+      codePoints = unicodes.map((u) => `0x${u}`)
+
+  return String.fromCodePoint(...codePoints)
+}
+
+export function sanitize(emoji) {
+  var { name, short_names, skin_tone, skin_variations, emoticons, unified, custom, imageUrl } = emoji,
+      id = emoji.id || short_names[0],
+      colons = `:${id}:`
+
+  if (custom) {
+    return {
+      id,
+      name,
+      colons,
+      emoticons,
+      custom,
+      imageUrl
+    }
+  }
+
+  if (skin_tone) {
+    colons += `:skin-tone-${skin_tone}:`
+  }
+
+  return {
+    id,
+    name,
+    colons,
+    emoticons,
+    unified: unified.toLowerCase(),
+    skin: skin_tone || (skin_variations ? 1 : null),
+    native: unifiedToNative(unified),
+  }
+}
+
